New and best version v2.0.1 (currently in BETA)
This is the only one of its kind include which actually adds new dialog styles to ShowPlayerDialog (the existing default samp native). The new dialogs are built on textdraws but have almost similar style to those of SAMP GUI.

One of the biggest benefit of this include is that it's of a very small size and gives a very minimal output size and all the dialog styles are in only one include.